[GStreamer][MSE] Rework marking end of append
In some cases current custom events can go thru the appsrc src pad
before the buffer itself, causing that the buffer is not processed. Now
we send a probe for buffers in the appsrc src pad and when a buffer goes
thru, we engage the process of marking the end of the append. The end of
the append is done at the main thread by sending an application message
thru the bus.